+++ h2o-core/src/main/java/water/DelegatingAuthenticator.java
@@ -11,7 +11,7 @@ import javax.servlet.ServletResponse;
 import javax.servlet.http.HttpServletRequest;
 
 /**
- * DelegatingAuthenticator dynamically switches between Form-based authentication
+ * Dynamically switches between Form-based authentication
  * and Basic Access authentication.
  * The decision is made based on user's "User-Agent". Browser clients will use Form based
  * authentication, all other clients will use basic auth.
